improve RBAC

+++ app/Http/Controllers/Admin/AdminRoleController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\AdminRole;
 use Facade\Ignition\Middleware\AddLogs;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class AdminRoleController extends Controller
 {
@@ -21,7 +22,7 @@ class AdminRoleController extends Controller
      */
     public function index()
     {
-        return AdminRole::paginate();
+        return AdminRole::with('actions')->get();
     }
 
     /**
@@ -37,31 +38,6 @@ class AdminRoleController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\AdminRole  $adminRole
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        return AdminRole::find($id);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\AdminRole  $adminRole
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        $role = AdminRole::find($id);
-        $role->name = $request->name;
-        $role->save();
-    }
-
-    /**
      * Remove the specified resource from storage.
      *
      * @param  \App\Models\AdminRole  $adminRole
@@ -70,12 +46,33 @@ class AdminRoleController extends Controller
     public function destroy($id)
     {
         $role = AdminRole::find($id);
-        $role->delete();
+
+        if ($role->name == 'admin')
+        {
+            return response()
+                ->json(['message' => "Not allowed to delete role 'admin'"], 422);
+        }
+
+        if (count($role->users) > 0)
+        {
+            return response()
+                ->json(['message' => "There are users assigned to this role.\nDetach first"], 422);
+        }
+
+        DB::transaction(function() use($role) {
+            $role->users()->detach();
+            $role->actions()->detach();
+            $role->delete();
+        });
     }
 
     public function assignActions(Request $request, $id)
     {
         $role = AdminRole::find($id);
+        if ($role->name == 'admin') {
+            return response()
+                ->json(['message' => "Role 'admin' no need to assign actions"], 422);
+        }
         $role->actions()->sync($request->action_ids);
     }
 }
